MODCLUSTER-607 Support floating-point numerals for decay factor
decayFactor and DEFAULT_DECAY_FACTOR type changed from int to float

getDecayFactor() return type and setDecayFactor() param type changed
from int to float

decayFactor type changed from int to float

MODCLUSTER-607 Support floating-point numerals for decay factor. Test
added

MODCLUSTER-607 Support floating-point numerals for decay factor. Use
default values

MODCLUSTER-607 Support floating-point numerals for decay factor. Format

Format

+++ core/src/main/java/org/jboss/modcluster/load/impl/DynamicLoadBalanceFactorProviderMBean.java
@@ -39,7 +39,7 @@ public interface DynamicLoadBalanceFactorProviderMBean {
      *
      * @return a positive integer
      */
-    int getDecayFactor();
+    float getDecayFactor();
 
     /**
      * Sets the number of historical load values to consider when calculating the load balance factor.
@@ -53,7 +53,7 @@ public interface DynamicLoadBalanceFactorProviderMBean {
      *
      * @param decayFactor the new decay factor
      */
-    void setDecayFactor(int decayFactor);
+    void setDecayFactor(float decayFactor);
 
     /**
      * Returns the load metrics registered with this provider
